# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.UserRequest:
    properties:
      id:
        type: integer
      is_admin:
        type: integer
      login:
        type: string
      name:
        type: string
      password_hash:
        type: string
    type: object
...
  repository.User:
    properties:
      id:
        type: integer
      isAdmin:
        type: integer
      login:
        type: string
      name:
        type: string
      passwordHash:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  title: Lab2 REST API
  version: "1.0"
paths:
...
  /users/all:
    get:
      description: Returns all user records (includes password hash fields as stored).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repository.User'
            type: array
        "405":
          description: Method not allowed
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: List users
      tags:
      - users
  /users/create:
    post:
      consumes:
      - application/json
      description: Admin only. Expects pre-hashed password in password_hash.
      parameters:
      - description: Bearer admin JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: login, name, password_hash, is_admin
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Validation error
          schema:
            type: string
        "401":
          description: Missing or invalid token
          schema:
            type: string
        "403":
          description: Not an admin
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - users
  /users/delete:
    delete:
      description: Admin only.
      parameters:
      - description: Bearer admin JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid id
          schema:
            type: string
        "401":
          description: Missing or invalid token
          schema:
            type: string
        "403":
          description: Not an admin
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - users
  /users/get:
    get:
      parameters:
      - description: User ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repository.User'
        "400":
          description: Invalid id or not found
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
      summary: Get user by ID
      tags:
      - users
  /users/update:
    put:
      consumes:
      - application/json
      description: Admin only.
      parameters:
      - description: Bearer admin JWT
        in: header
        name: Authorization
        required: true
        type: string
      - description: id, name, password_hash, is_admin
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Validation error
          schema:
            type: string
        "401":
          description: Missing or invalid token
          schema:
            type: string
        "403":
          description: Not an admin
          schema:
            type: string
        "405":
          description: Method not allowed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - users
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"